# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KoaPOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pos-registers:
    get:
      operationId: listPosRegisters
      summary: List POS registers
      tags: [POS]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosRegisterListResponse" }
    post:
      operationId: createPosRegister
      summary: Create POS register
      tags: [POS]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PosRegisterInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosRegister" }
  /pos-registers/{id}:
    patch:
      operationId: updatePosRegister
      summary: Update POS register
      tags: [POS]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PosRegisterInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosRegister" }
    delete:
      operationId: deletePosRegister
      summary: Delete POS register
      tags: [POS]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /pos-settings:
    get:
      operationId: getPosSettings
      summary: Get POS settings
      tags: [POS]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosSettings" }
    put:
      operationId: upsertPosSettings
      summary: Upsert POS settings
      tags: [POS]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PosSettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosSettings" }
...
  /qr-codes:
    get:
      operationId: listQrCodes
      summary: List QR codes
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrCodeListResponse" }
    post:
      operationId: createQrCode
      summary: Create QR code
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QrCodeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrCode" }
  /qr-codes/{id}:
    delete:
      operationId: deleteQrCode
      summary: Delete QR code
      tags: [Marketing]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /qr-settings:
    get:
      operationId: getQrSettings
      summary: Get QR settings
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrSettings" }
    put:
      operationId: upsertQrSettings
      summary: Upsert QR settings
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QrSettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrSettings" }
  /qr-saved-templates:
    get:
      operationId: listQrSavedTemplates
      summary: List saved QR templates
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrSavedTemplateListResponse" }
    post:
      operationId: createQrSavedTemplate
      summary: Save QR template
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QrSavedTemplateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QrSavedTemplate" }
  /qr-saved-templates/{id}:
    delete:
      operationId: deleteQrSavedTemplate
      summary: Delete saved QR template
      tags: [Marketing]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /landing-pages:
    get:
      operationId: listLandingPages
      summary: List landing pages
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LandingPageListResponse" }
    post:
      operationId: createLandingPage
      summary: Create landing page
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LandingPageInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LandingPage" }
  /landing-pages/{id}:
    patch:
      operationId: updateLandingPage
      summary: Update landing page
      tags: [Marketing]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LandingPageInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LandingPage" }
    delete:
      operationId: deleteLandingPage
      summary: Delete landing page
      tags: [Marketing]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /shortlinks:
    get:
      operationId: listShortlinks
      summary: List shortlinks
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShortlinkListResponse" }
    post:
      operationId: createShortlink
      summary: Create shortlink
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShortlinkInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shortlink" }
  /shortlinks/{id}:
    delete:
      operationId: deleteShortlink
      summary: Delete shortlink
      tags: [Marketing]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /shortlink-settings:
    get:
      operationId: getShortlinkSettings
      summary: Get shortlink settings
      tags: [Marketing]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShortlinkSettings" }
    put:
      operationId: upsertShortlinkSettings
      summary: Upsert shortlink settings
      tags: [Marketing]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShortlinkSettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShortlinkSettings" }
...
components:
  schemas:
...
    PosRegister:
      type: object
      required: [id, merchantId, registerId, name, type, staffName, staffEmail, posCameraEnabled]
      properties:
        id:                { type: integer }
        merchantId:        { type: integer }
        registerId:        { type: string }
        name:              { type: string }
        type:              { type: string }
        staffName:         { type: string }
        staffEmail:        { type: string }
        posCameraEnabled:  { type: string }
        posCameraDeviceId: { type: string, nullable: true }
    PosRegisterInput:
      type: object
      required: [registerId, name, type]
      properties:
        registerId:        { type: string }
        name:              { type: string }
        type:              { type: string }
        staffName:         { type: string }
        staffEmail:        { type: string }
        posCameraEnabled:  { type: string }
        posCameraDeviceId: { type: string, nullable: true }
    PosRegisterListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PosRegister" }
        total: { type: integer }
    PosSettings:
      type: object
      required: [id, merchantId, enabledPaymentMethods, enabledIntegrationPayments, gridColumns, gridTileSize, gridShowPrices, gridShowStockBadges, gridCartPosition, forceStaffLogin, staffLoginMessage, activeRegisterId, hardwareConfig, enabledShortcuts]
      properties:
        id:                          { type: integer }
        merchantId:                  { type: integer }
        enabledPaymentMethods:       { type: string }
        enabledIntegrationPayments:  { type: string }
        gridColumns:                 { type: integer }
        gridTileSize:                { type: string }
        gridShowPrices:              { type: string }
        gridShowStockBadges:         { type: string }
        gridCartPosition:            { type: string }
        forceStaffLogin:             { type: string }
        staffLoginMessage:           { type: string }
        activeRegisterId:            { type: string }
        hardwareConfig:              { type: string }
        enabledShortcuts:            { type: string }
    PosSettingsInput:
      type: object
      properties:
        enabledPaymentMethods:       { type: string }
        enabledIntegrationPayments:  { type: string }
        gridColumns:                 { type: integer }
        gridTileSize:                { type: string }
        gridShowPrices:              { type: string }
        gridShowStockBadges:         { type: string }
        gridCartPosition:            { type: string }
        forceStaffLogin:             { type: string }
        staffLoginMessage:           { type: string }
        activeRegisterId:            { type: string }
        hardwareConfig:              { type: string }
        enabledShortcuts:            { type: string }
...
    QrCode:
      type: object
      required: [id, merchantId, entryId, label, url, qrType, content, settings, createdAt]
      properties:
        id:          { type: integer }
        merchantId:  { type: integer }
        entryId:     { type: string }
        label:       { type: string }
        url:         { type: string }
        qrType:      { type: string }
        content:     { type: string }
        settings:    { type: string }
        createdAt:   { type: string, format: date-time }
    QrCodeInput:
      type: object
      required: [entryId, label]
      properties:
        entryId:     { type: string }
        label:       { type: string }
        url:         { type: string }
        qrType:      { type: string }
        content:     { type: string }
        settings:    { type: string }
    QrCodeListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QrCode" }
        total: { type: integer }
    QrSettings:
      type: object
      required: [id, merchantId, patternColor, eyeColor, eyeDotColor, bgColor, dotStyle, cornerSquareStyle, cornerDotStyle, template, size, level, logoUrl]
      properties:
        id:                  { type: integer }
        merchantId:          { type: integer }
        patternColor:        { type: string }
        eyeColor:            { type: string }
        eyeDotColor:         { type: string }
        bgColor:             { type: string }
        dotStyle:            { type: string }
        cornerSquareStyle:   { type: string }
        cornerDotStyle:      { type: string }
        template:            { type: string }
        size:                { type: integer }
        level:               { type: string }
        logoUrl:             { type: string }
    QrSettingsInput:
      type: object
      properties:
        patternColor:        { type: string }
        eyeColor:            { type: string }
        eyeDotColor:         { type: string }
        bgColor:             { type: string }
        dotStyle:            { type: string }
        cornerSquareStyle:   { type: string }
        cornerDotStyle:      { type: string }
        template:            { type: string }
        size:                { type: integer }
        level:               { type: string }
        logoUrl:             { type: string }
    QrSavedTemplate:
      type: object
      required: [id, merchantId, templateId, name, settings, createdAt]
      properties:
        id:          { type: integer }
        merchantId:  { type: integer }
        templateId:  { type: string }
        name:        { type: string }
        settings:    { type: string }
        createdAt:   { type: string, format: date-time }
    QrSavedTemplateInput:
      type: object
      required: [templateId, name, settings]
      properties:
        templateId:  { type: string }
        name:        { type: string }
        settings:    { type: string }
    QrSavedTemplateListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/QrSavedTemplate" }
        total: { type: integer }
...
    LandingPage:
      type: object
      required: [id, merchantId, pageId, slug, title, links, createdAt, updatedAt]
      properties:
        id:            { type: integer }
        merchantId:    { type: integer }
        pageId:        { type: string }
        slug:          { type: string }
        title:         { type: string }
        subtitle:      { type: string }
        bio:           { type: string }
        profileImage:  { type: string }
        bgType:        { type: string }
        bgColor:       { type: string }
        bgFrom:        { type: string }
        bgTo:          { type: string }
        bgDir:         { type: string }
        bgImage:       { type: string }
        btnStyle:      { type: string }
        btnVariant:    { type: string }
        btnBg:         { type: string }
        btnText:       { type: string }
        btnBorder:     { type: string }
        textColor:     { type: string }
        font:          { type: string }
        links:         { type: string }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    LandingPageInput:
      type: object
      required: [pageId, slug, title]
      properties:
        pageId:        { type: string }
        slug:          { type: string }
        title:         { type: string }
        subtitle:      { type: string }
        bio:           { type: string }
        profileImage:  { type: string }
        bgType:        { type: string }
        bgColor:       { type: string }
        bgFrom:        { type: string }
        bgTo:          { type: string }
        bgDir:         { type: string }
        bgImage:       { type: string }
        btnStyle:      { type: string }
        btnVariant:    { type: string }
        btnBg:         { type: string }
        btnText:       { type: string }
        btnBorder:     { type: string }
        textColor:     { type: string }
        font:          { type: string }
        links:         { type: string }
    LandingPageListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LandingPage" }
        total: { type: integer }
    Shortlink:
      type: object
      required: [id, merchantId, linkId, label, longUrl, slug, baseDomain, clicks, tags, createdAt]
      properties:
        id:          { type: integer }
        merchantId:  { type: integer }
        linkId:      { type: string }
        label:       { type: string }
        longUrl:     { type: string }
        slug:        { type: string }
        baseDomain:  { type: string }
        clicks:      { type: integer }
        tags:        { type: string }
        createdAt:   { type: string, format: date-time }
    ShortlinkInput:
      type: object
      required: [linkId, label, longUrl, slug]
      properties:
        linkId:      { type: string }
        label:       { type: string }
        longUrl:     { type: string }
        slug:        { type: string }
        baseDomain:  { type: string }
        tags:        { type: string }
    ShortlinkListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Shortlink" }
        total: { type: integer }
    ShortlinkSettings:
      type: object
      required: [id, merchantId, baseDomain, prefix]
      properties:
        id:          { type: integer }
        merchantId:  { type: integer }
        baseDomain:  { type: string }
        prefix:      { type: string }
    ShortlinkSettingsInput:
      type: object
      properties:
        baseDomain:  { type: string }
        prefix:      { type: string }
...